backend/requirements.txt
fastapi
uvicorn
vcfpy
python-multipart
python-dotenv
google-generativeai
